Change HTTP status codes and fix bug
Numbers were used in the code for response HTTP status codes, so these
were changed to constant variable names found in the HttpStatus.util.js.

There was a bug within the changeUserPermissions function in
User.controller.js. If there was an error detected in one of the
user's details provided, then it would result in this error being
thrown:

Error [ERR_HTTP_HEADERS_SENT]: Cannot set headers after they are sent to
the client
    at new NodeError (node:internal/errors:405:5)
    at ServerResponse.setHeader (node:_http_outgoing:652:11)
    at ServerResponse.header

This was due to the way error responses were being handled in the loop.
Changed the way errors are handled so that all error messages are added
to an array first, then returned to the user after the loop finishes
executing.